requirements.txt
btclib==2020.12.19
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.2.3
dataclasses-json==0.5.2
ecdsa==0.14.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
locust==1.4.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
msgpack==1.0.2
mypy-extensions==0.4.3
psutil==5.8.0
pyasn1==0.4.8
python-jose==3.2.0
pyzmq==22.0.3
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7
six==1.15.0
stringcase==1.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
